+++ Migratable/Interfaces/IMigrator.cs
@@ -6,19 +6,19 @@ namespace Migratable.Interfaces
     public interface IMigrator
     {
         /// <summary>Loads all the migrations in the given folder path.</summary>
-        SortedList<long, Migration> LoadMigrations(string folderPath);
+        SortedList<int, Migration> LoadMigrations(string folderPath);
 
         /// <summary>Register for version movement notifications.</summary>
         void SetNotifier(INotifier notifier);
 
         /// <summary>Get the current version.</summary>
-        long GetVersion();
+        int GetVersion();
         /// <summary>Roll to the specified version.</summary>
-        void SetVersion(long targetVersion);
+        void SetVersion(int targetVersion);
 
         /// <summary>Specifically roll forward to the given version.</summary>
-        void RollForward(long targetVersion);
+        void RollForward(int targetVersion);
         /// <summary>Specifically roll backward to the given version.</summary>
-        void RollBackward(long targetVersion);
+        void RollBackward(int targetVersion);
     }
 }
